Update PR template
Problem: GitHub has special keywords to close an issue when a PR
is merged, but our PR template doesn't encourage their usage.
Most commonly a PR closes an issue, so it makes sense to use a
corresponding keyword.

Solution: update "related issue(s)" section to contain "Resolves"
keyword. Also use just `#` instead of full link to an issue because
it is shorter and more resistant to changes like movement of a repo.