Update CI badge
Problem: CI badge shows status for latest build which is quite useless
information because last build can be in some WIP PR where it is even
not supposed to pass.

Solution: Add `?branch=master` to `.svg` link, it should show status
of the latest `master` build which is more useful.
Also remove `?branch=master` from the link that will be open on click
because it seems to have no effect there.