[#85] Add tezos-codec packeg
Problem: Apart from all binaries that we're currently packaging, there is
tezos-codec binary.We should find out what is the purpose of this binary
and package it in case it's useful.

Solution: tezos-codec is a binary to decode and encode JSON. Added the
ability to create static binaries using docker and nyx. It is also possible
to build packages using docker.