[#109] Build tezos-sandbox static binary
Problem: There is a program called tezos-sandbox that can be helpful
for setting up and testing various scenarios with the blockchain.
Also, it's not built by default, however, it's still usefult for the
developers

Solution: Build and bundle static tezos-sandbox binary.

Suddenly, there is no opam package for tezos-sandbox, thus we cannot
provide native Ubuntu and Fedora packages with tezos-sandbox.