[#103] Add formulas sanity check
Problem: Installing all formulas in CI will consume a lot of time
(~12-15 minutes per formula), so we need another way to check that
formulas are sane enough.

Solution: Run 'brew info' and 'brew fetch -s', thus we'll be sure that
there are no syntax errors in the formulas and the Tezos sources can be
fetched.