[#132] Refactor packages creation once again
Problem: We need to add one more package, which will be produced in a
bit different way. Currently all packages are based on the bundled opam
package and the existing approach couldn't be easily adopter for
packaging anything different.

Solution: Move packaging-related files generation to the AbstractPackage
class, so that it's possible to have packages, whose packaging-related files
are created differently.