[#132] Add tezos-sapling-params package
Problem: Sapling-params is a hard runtime dependency for some of the
Tezos binaries since v8.0. It's inconvenient to download them manually.

Solution: Add one more native package with required params files.
This package will be used in the other packages dependencies, thus it's
supposed to be installed automatically along with the Tezos binaries.