Use ~/.tezos-client as a daemons data dir
Problem: All commands within Tezos services are run under 'tezos' user.
It's quite inconvenient to provide '-d /var/lib/tezos/client' to the
'tezos-client' commands when we want to work with the keys that are
supposed to be known to the daemons by default.

Solution: Use default tezos-client data directory (which is located in
~/.tezos-client) as a default data directory for daemon services.