[#130] Don't inherit abstract Tezos class
Problem: Currently all our formulas actually consist of two files:
abstract 'Tezos.rb' which is inherited by the rest of the formulas.
Such an approach make it impossible to add services to the formulas and
also use formulas as a dependencies. This is happens due to the fact
that abstract 'Tezos.rb' isn't included along with the 'Tezos*.rb' and
thus brew fails when it's trying to scan the installed formula.

Solution: Don't use the abstract class and copypaste common parts.