Don't depend on rust brew formula
Problem: Tezos requires rust version to be specified explicitely
in case a version different from the default one (1.44.0) is used. Previously, we
depended on rust brew formulas, but the rust version in it can be
changed at any time, which means that our formulas are quite unstable.

Solution: Depend on rustup-init and fetch rust compiler via
`rustup-init --default-toolchain 1.44.0`.