[#156] Move node identity generation to prestart
Problem: Currently node identity generation is done during `tezos-node
run`. This process takes some significatnt amount of time and thus baker
and endorser daemons are failing with connection timeout to the node.

Solution: Move identity generation to `ExecStartPre`, so that identity
is generated before the node and its dependencies are started.