Mention mirrored taps in the README
Problem: We'd like to have an ability to install both latest stable
version and latest release candidate brew formulas on the macOS.
However, brew tap uses master branch of the repo by default and it
cannot be changed.

Solution: Create two mirrors of the tezos-packaging repo, one mirror for
the latest stable release and one mirror for the latest release
candidate release, thus users may use one of these repos as a brew tap
and install desired stable or release candidate formulas. Mention this
capability in the README.