[#156] Unify BAKER_ACCOUNT and ENDORSER_ACCOUNT
Problem: BAKER_ACCOUNT and ENDORSER_ACCOUNT should point at the same
alias in most of the cases, so having to specify it twice can be
annoying. Though, there is still a usecase when the endorser can be used
alone without baker and local-node, but this use-case is quite rare.

Solution: Replace BAKER_ACCOUNT and ENDORSER_ACCOUNT with BAKER_ADDRESS_ALIAS,
so that this new environment variable can be specified only once in the
'/etc/default/tezos-baking-<network>'.