[#156] Use 'baker' as default BAKER_ADDRESS_ALIAS
Problem: User still has to specify alias in the config file and import
the key with this alias under the 'tezos' user. However, we'd like to
reduce the amount of the manual setup as much as possible.

Solution: Use 'baker' as a default 'BAKER_ADDRESS_ALIAS' in
'/etc/default/tezos-baking-<network>', thus user will only have to
import key with the 'baker' alias.