[#175] Extract meta to a dedicated module
Problem: Meta information for the packages: version, release number,
epoches, maintainer info is fetched from the json file and hardcoded in
the packages.py. It's quite nontrivial to find where it's defined in
case when changes to it are required.

Solution: Extract meta information to a dedicated module meta.py, all
meta information is now stored in the PackagesMeta class, instance of
this class is used in all packages to define meta infromation.