[#195] Dynamically generate pipeline via python
Problem: Some of the ci steps take huge amount of time. We run them on
any changes even when the changes don't affect any of it.

Solution: Make it possible to provide regexes, which will define paths,
whose changes will trigger step to run. Pipeline from the yaml file
is loaded to the python dictionary, this dict is later filtered
basing on the output of 'git diff origin/master...HEAD --name-only'
or 'git diff --name-only HEAD HEAD~1' in case we're in master branch.
The files from diff list are matched agains the regexes from the
pipeline config. If step doesn't match any of them, then it isn't
included in the resulting generated pipeline config.

To generate pipeline config locally run:
```
./.buildkite/generate-pipeline.sh
```