[#201] Update README and the baking guide
Problem: we have added a CLI setup wizard for tezos baking, but there
is no mention of it in the documentation.

Solution: update README.md and the baking guide for Ubuntu to mention
the wizard and how to use it. Additionally, fix and/or update some places
in the baking guide with new information about supported protocols and workflow.