[#226] Add PR template for releases
Problem: The process of creating a new release in Tezos packaging
consists of many steps and it's quite important to not miss any of them.
Since the process of creating a new release usually consist of multiple
PRs, it'll be nice to have a place where we can track the things that
was done and needs to be done in the nearest future.

Solution: Add PR template that is supposed to be used when we're
intending to create a new release. This PR will be used to track the
progress of the new release.