Add nix experssions for publishing utilities
Problem: 'dput' and 'copr-cli' are needed in order to submit Ubuntu and
Fedora package to the Launchpad and Copr. However, this utilities aren't
presented in the nixpkgs at the moment, as a result it's quite
inconvenient to use them on NixOS.

Solution: Provide expressions for these utilities, add check to the CI
that they can be successfully built.