[#242] Fix baking wizard failing on Python 3.6
Problem: in the baking wizard we use the subrocess lib for calling
external processes like `systemctl`. The API changed slightly from
Python 3.6 to 3.7, we need to support Ubuntu 18.04 which by default
has the former, and right now the wizard consistently crashes on it.

Solution: added the differentiation for the culprit method
subprocess.run for whether or not to use the capture_output parameter
based on the Python version running the wizard.