[#251] Split main README into multiple docs
Problem: At the moment main README.md is quite large and it's hard to
navigate through it, even though it has table of contents.
Usually, users are interested in only one way of distribution and it's
not convenient for them to skim through the other distribution options.

Solution: Split README into multiple docs:
* README itself that points to the other docs.
* Ubuntu-related instructions.
* Fedora-related instructions.
* macOS-related instructions.
* Doc about systemd services.
* Doc about launchd services.