[#256] Improve ledger import prompt
Problem: Currently it's possible to choose one of the four ledger keys
provided by tezos-client by default or type a ledger url manually.
However, we doesn't properly validate that ledger url is valid.
Instead we should provide a way to choose a derivation path that will be
used with the current ledger device.

Solution: Make it possible to provide custom derivation path instead of
the whole ledger url. Once the derivation is provided it appears in the
list of possible ledger urls to choose from. As an addition, an option
to go back to the baker key import process beginning is added.