[#251] Turn README header links into a bullet-point list
Problem: the header installation links take up a fairly large
amount of space and as a result they aren't easily recognizable as
links. Additionally the subtext is mostly repetitive and doesn't
really add much to it.

Solution: simplify the section by turning all sub-sections into a
bullet-point list, with the exception of static linux binaries.