[#251] Remove list of binaries from the README
Problem: at the very beginning of our README we have a fairly long
list of the binaries provided, which is mostly just distracting.
Moreover this list as well as some of the additional info can be
instead found on the official docs, among other explanations, that
we have to link to either way.

Solution: simplify the opening paragraph, by changing it to a brief
introducion to the repo followed by a link to the Tezos website.