[#251] Simplify and put in front automated setup
Problem: our most well supported package use is the automated
installation and setup on Ubuntu, however this is hidden after other
installation links (including one for Ubuntu), whereas we should have
it in appealing spot on the README and with a simpler intro.

Solution: move the dedicated section to the top and show the shortest
command list to start a setup.