[#276] Pass one env var to suppress testnet warning
Problem: to suppress the "This is NOT the Tezos Mainnet" warning,
we currently propagate the whole environment through sudo. This
leads to HOME being preserved under 'sudo -u tezos', and
tezos-client creating its directory in the original user HOME.

Solution: changed all the tezos-client calls to directly pass the
one environment variable needed for this,
TEZOS_CLIENT_UNSAFE_DISABLE_DISCLAIMER, instead of the whole env.