[#284] Add script automating bottle hash insertion
Problem: every time after a new release, we build brew bottles,
and their SHA256 hashes should be added to the brew formulae.
Right now we do this manually, but it's a tedious and error-prone
process that would be nice to automate.

Solution: added a bash script that can do this. It takes a directory
where the bottle.tar.gz files are placed as its argument and inserts
the computed hash for the respective OS into the right formula.
It is supposed to be run from the base directory (tezos-packaging).