[#284] Add tests for bottle hash insertion script
Problem: we have added a new script that automates inserting brew
bottle hashes, but there are no tests for it.

Solution: added another bash script that generates a dummy formula
and some dummy bottles and checks that the result of running
bottle-hashes.sh is as expected. Also added this new step to buildkite.