[#286] Automate bottle hash PR opening
Problem: we need to automate brew bottles-related processes further.
Right now we automatically build and attach the bottles to releases, and
we have a script that computes and inserts their hashes to the respective
formulae. It's a natural next step to open the PRs from the CI.

Solution: added a step to each pipeline that should sync the hashes
in a common branch, and then open a PR automatically.