Refactor Tezos static build in docker
Problem: It's quite inconvenient to maintain a patch file with compiler
options that are required for building statically linked binaries.

Solution: Use static dune profile instead which was provided by the
Octez developers some time ago. This profile can be used for all
binaries except for the tezos-sandbox, so we preserve a patch file for
it for now.

Additionally bump used RUST version, since alpine now has version that
is incompatible with Octez by default.