[#288] Add script for release PR creation
Problem: We want to automate update routine in tezos-packaging.
One of things that is currently done manually but can be automated is
initial release PR creation.

Solution: Add a script that will fetch the latest release tag in the
upstream repository, compare it with the upstream revision in our repo
and create a PR with initial required changes that include the
following updates:
* Tezos sources revision
* opam-repository revision
* brew formulae updates (bump version and remove old hashes).