[#314] Add automatic mirror repositories updates
Problem: we have automated almost the whole release making process.
However, we use two mirror repositories for macOS formulae, and this
step is still performed manually.

Solution: added two steps to the Buildkite pipeline that run on master
when specific merge commits are made: either the formulae have been
updated for the new Octez version, or we have built the bottles and
consequently added their hashes to the formulae.