[#334] Auto create release upon release PR merge
Problem: we have automated almost the whole release making process
except for actually publishing the release from autoreleases. Currently,
after the release PR is merged and a auto pre-release made, we have to
manually push a tag and publish a release with a correct name.

This has already lead to maintainers forgetting to publish a release,
which led to some CI issues. We would like to mitigate this by
automating this process as well.

Solution: reworked the auto-release making process to actually make full
releases/pre-releases if we merged a release PR, or still make an
'auto-release' if not.