[#314] Run mirror updates after autorelease
Problem: right now, the mirrors are updated right as the CI build is
triggered. If we wait until after the (pre-)release is automatically
published, we get more info, e.g. tags, in the mirror repo as well.

Solution: added dependencies on the 'auto-release' CI step for the
mirror updates step.