[#323] Add dependencies to tezos-sandbox formula
Problem: there is a known issue with using tezos-sandbox on MacOS, which is that it
depends on coreutils and util-linux. These are not installed upon running
'brew install tezos-sandbox'.
Additionally, homebrew installs all tools that conflict with MacOS ones in name with
the 'g' prefix, meaning the user needs to manipulate $PATH to use these dependencies.

Solution: added coreutils and util-linux as dependencies to the formula.
As brew does not allow for post-setup user configuration manipulation, $PATH cannot be
easily modified to include gnubin. We use the paradigmatic homebrew way and output the
command in the caveats section for the user to run manually.