[Chore] Refactor bottle-building GA config
Problem: when building bottles using GitHub Actions, we had a whole job
copy-pasted to support running on both Catalina and Big Sur what was
essentially the same process. This is wasteful.

Solution: changed that workflow to use matrix, allowing us to define
what OS to run on and the corresponding OS name to use in scripts.