[Chore] Hotfix matrix value propagation in GA
Problem: in the GitHub Actions for building brew bottles, though we have defined
the OS names in the strategy matrix to be in quotes, these quotes are seemingly
lost when substituting the matrix value into the script call. This causes the
commit message for the Big Sur run have "Big" instead of "Big Sur".

Solution: added explicit quotes to the OS name argument in the script call.