[#364] Support archive nodes from snapshot import
Problem: there is a use-case for being able to reconstruct storage
to run an archive node when importing a full snapshot. The Setup Wizard
at the moment doesn't support running archive nodes in any capacity
except by manually changing the code, which isn't very convenient.

Solution: if the snapshot the user wants to import is in full history
mode, inquire whether the storage should be reconstructed to yield
an archive node.