[#366] Add history mode support in setup wizard
Problem: currently, the only history mode support present in the Tezos
Setup Wizard is the ability to reconstruct an archive mode when importing
a full snapshot, and to import either a full or a rolling snapshot. This
means we have no history mode support for syncing the node from scratch,
and the default mode is 'full', which might not be reasonable.

Solution: added a new step in the wizard, asking about the preferred history
mode in case the user skips snapshot import.