[#361] Rework Buildkite bottle building
Problem: the internet in our HQ where the Big Sur arm64 Mac resides
is sometimes unstable, which leads to some bottles sometimes failing
to build. Rebuilding is non-straightforward, since we must rebuild the
whole batch. This takes a lot of time and the success is still not
guaranteed.

Solution: reworked the tags pipeline to be in two parts: the raw
pipeline-for-tags.yml and a bottle-building part, generated dynamically
to avoid tons of copy-paste, that is appended to the raw part. Note: this
requires us to change that pipeline's settings using Buildkite's UI to
upload the generated .yml.