[Chore] Fix GA bottle workflow
Problem: syncing bottle hashes fails because it requires GITHUB_TOKEN
in the env, which was not provided. Also, many bottles for Big Sur won't
build because there is a typo in the 'already-built' checker that will
mistake arm64_big_sur bottles for big_sur ones.

Solution: added an env parameter to 'download bottles from release' steps,
also fixed the typo in the bottle checker.