[Chore] Fix tezos-baking build on Fedora
Problem: tezos-baking had failed to build on any Fedora version or
architecture for months, because the startup scripts, tezos-baking-start
and tezos-baking-prestart, were expected but not found in the build dir.

Solution: maybe fixed it by explicitly creating /usr/bin/ before
installing the startup scripts in that directory. Also changed all
docker scripts to be in 755 mode for consistency.