[Chore] Add udev rules for ledger devices in WSL
Problem: in #199 we excluded WSL in the post-installation
script from installing the udev rules because the reloading
resulted in a failure.
However, the udev rules for ledger devices can be used and
there is a workaround for the reloading issue, so we should
make use of it.

Solution: modify the post-install script to install udev
rules in WSL by using the service restart workaround.