[#385] Document workaround for zcash-params on M1 Macs
Problem: 'HOMEBREW_PREFIX' has changed on Macs with Apple CPUs. As a
result, Octez binaries are unable to find zcash-params installed by
'tezos-sapling-params' formula.

Solution: It's impossible to fix this issue properly from the brew formula
side. So we document a workaround instead. Additionally, we've
submitted a patch to the upstream that targets to resolve this issue
and it is hopefully going to be included in the next release.