[#401] Don't require tezos-baking-mainnet.service to vote
Problem: currently, the voting wizard only supports systems which
have baking through systemd services set up. This is not the actual
minimal requirement, which is access to a ledger via tezos-client.

Solution: only check the service for baking info if it's running,
otherwise collect the necessary information from the user.