[Chore] Add baker secret key option to the baker module
Problem: Currently, the only way to setup baker is to import key to
the required data directory and with the required alias manually.
However, this isn't very convenient for automatic deployments.

Solution: Allow specifying baker secret key as a baker module option.
This key will be imported with a given alias before launching the baker
daemon.