[#418] Reduce bottle building in BK to one step
Problem: currently, after each release, we partially generate the
Buildkite pipeline to handle building brew bottles for formulae
relevant to the release. This script has been extended since its
conception and is becoming difficult to maintain.

Solution: reworked the BK tags pipeline to include a single step
for each OS, calling an external script that filters the formulae
to the list of those that have been changed, and builds and uploads
them one-by-one.