[Chore] Use jakartanet instead of its config URL
Problem: in v13.0-rc1, there was no predefined jakartanet option for
tezos-node, so we have used a URL to fetch a config. v13.0 introduces
support for --network jakartanet, so we should switch to that option too.

Solution: switched to jakartanet option wherever the URL was used, added
nix tests for jakartanet.