[Chore] Fix 'update-tezos.sh'
Problem: when finding out 'opam_repository_branch', refnames used to
be shown as 'refs/remotes/origin/*' and not it's 'refs/heads/*'.
This lead to an incorrect nix/nix/source.json update.

Solution: changed 'git for-each-ref' to use a more general path,
assumed branch names are not going to have slashes in them and
started splitting the rightmost part.