[Chore] Fix autorelease triggering invalid CI runs
Problem: when publishing an auto-release/pre-release, we have to repush
the 'auto-release' tag, as it always exists. However, currently we repush
all tags in 'scripts/autorelease.sh', and sometimes it can lead to
unwanted/unexpected rebuilds of CI for old tags, re-published native
packages that we don't need, and other problems.

Solution: changed the script to only repush the auto-release/pre-release
tag.