[#441] Use xtz-shots.io snapshot metadata
Problem: in Tezos Setup Wizard, we allow to automatically download
a suitable snapshot from xtz-shots.io. The links have a set pattern
that we can predict and reuse, but it's not necessarily a stable
solution. xtz-shots.io recently added a metadata JSON to their site,
and using it to get snapshot URLs would be more robust.

Solution: switched the snapshot download step to fetching snapshot
metadata from base.json and choosing the most recent available entry
to offer as options.