[Chore] Update documentation on scripts
Problem: there are some scripts whose usage is mentioned in the
documentation, but they are hardly ever used or tested.
As a consequence, some of these examples need to be rechecked.

Solution: update some of the documentation and leave comments to
hopefully avoid this in the future.
Additionally adds a quick CI step to check the script that's most
likely less used.