[#464] Query toggle vote in setup wizard
Problem: in anticipation of jakarta on mainnet, we need to properly
support the liquidity baking toggle vote option that is now mandatory.
The Setup Wizard needs to ask for the user's preference.

Solution: added a step that queries the user's preference. There seems
to be no good way to provide this option for our systemd service, so
we modify the service's config file on each wizard invocation.