[Chore] Bump release number for jakartanet switch prep
Problem: we're going to need to make a new release in the scope of
the same upstream release, so we need to bump various versions.

Solution: bumped the release number, added the letter version because
of changes to the setup wizard, also added jakarta to the list of
mainnet protos.