[#484] Improve the update message about liquidity toggle vote
Problem: On updating (or installing) tezos-baking's latest version
a message is printed about the liquidity toggle vote option that is
mandatory since jakarta, but this does not tell the user what needs
to be done nor how.

Solution: Improve the update message by making it stand out more and
adding info about the default value and using tezos-setup-wizard to
change this option.