[Chore] Fix interpolation bug in autorelease.sh
Problem: in the autorelease.sh script the interpolation of an empty
variable results in two consecutive spaces, which is incorrectly
interpreted as an empty argument by the 'gh' command.

Solution: remove the double quotes to solve the issue and add a
shellcheck ignore to avoid marking this as incorrect.
Note: this partially reverts d7732417d282196f55cc5847b6f4d8dd466c9781