[Chore] Fail in CI when bottles were not built or uploaded
Problem: currently in case a bottle fails to be built, or uploaded,
an error message is printed, but the exit code remains 0.
This is a problem because one can miss this easily in the logs
and also because the successful step cannot be retried on CI.

Solution: return a non-0 exit code when any of the bottles failed to
build or was not uploaded.
Additionally add automated retries of the bottle building steps in CI.