[Chore] Always remove the initial 'v' from commonMeta.version
Problem: currently the 'commonMeta.version' as defined in
tezos-release.nix will drop the initial "v" from the 'branchName'
only when this starts with "refs/tags/", but in some cases this
part is missing.

Solution: modify the definition to drop the initial "v" whether
or not there is a "refs/tags/" at the start of the 'branchName'.