[#517] Use know data dir env vars without CLI args
Problem: we have several scripts and wrapper that
use env vars to decide what CLI argument gets passed
to Octez binaries.
However now that we use env vars that are known by
the binaries already, we don't need to use the args.

Solution: avoid this unnecessary conversion and use
the env vars directly.