[#517] Replace NODE_RPC_ENDPOINT with SCHEME and ADDR
Problem: There is currently duplication in the env var
definitions of the node's address and the env var of
baking's node endpoint.
This causes issues in updating as both env vars need to
be modified if one changes.

Solution: Remove NODE_RPC_ENDPOINT and replace it with
NODE_RPC_ADDR and NODE_RPC_SCHEME.
Remove the NODE_RPC_ADDR part from the baking service
and rely on the env var set in the node instead.