[Chore] Add lima baker to mainnet and ghostnet
Problem: this is the first stable release with full lima
support. If accepted the lima protocol will be activated
in both ghostnet and mainnet. Since an idle baker won't
consume many resources we can activate them already (to
avoid making a release specifically for that in the
future).

Solution: add lima baker to mainnet and ghostnet.