[#572] Use block hash from snapshot metadata in setup wizard
Problem: octez-node has a 'block' option when importing a
snapshot that can be used to ensure that the block that
is imported is the one matching the given hash.
We can collect the block hash from the metadata (when one
is available), but the setup wizard doesn't use this info.

Solution: get the block hash when fetching snapshots from
metadata and use it during import in the setup wizard.