[#571] Handle empty or missing node data dirs in setup wizard
Problem: it's possible for the tezos node data directory to be
empty or non-existent when the setup wizard is launched, most
notably if a custom one has been set.
However the wizard currently expects the node data dir to both
exist and contain a config (and possibly more), causing
'octez-node' to fail when this is not the case.

Solution: modify the wizard to handle missing and empty node data
directory, by creating and configuring them if necessary.