[#571] Make clear which node data dir is used by the setup wizard
Problem: in case a custom data directory is specified and something
goes wrong, it can be hard to figure out if the wizard is operating
on the right path or not.
This makes debugging harder and undermines the confidence that the
custom data directory option is in fact being used.

Solution: Change the related messages in the wizard so that whenever
the node data dir is manipulated, its path is printed on screen.