[Chore] Remove support for kathmandu
Problem: kathmandunet has been abandoned and is no longer producing
blocks. Ghostnet and Mainnet have also switched away to lima.
As such, there is currently no network running the kathmandu protocol
and we should update our docs and definitions accordingly.

Solution: remove any mention of kathmandu from the documents, remove
it from the list of active protocols, delete binaries and packages
that are specific to it and move remaining uses to lima(net).