[#588] Remove unnecessary conversion of snapshot metadata
Problem: the snapshot metadata provided by xtz-shots.io used
to contain a "block_height" value as a string, but it's now
a number.
The tezos-setup wizard converts this value to a number, but
that's no longer necessary

Solution: remove the unnecessary conversion of "block_height"
from the snapshot metadata to a number in tezos-setup.