[#606] Restart baking service after voting if necessary
Problem: voting (including with our wizard) using a ledger device
requires switching to the wallet app. If there is a local baking setup
running this also means that this will be interrupted until the Baking
app is opened on the ledger again.
Moreover, because of tezos/#4486 this is no longer sufficient and the
baker process needs to be restarted.

Solution: modify the voting wizard so that if one votes with a ledger
that is also used for a local baking setup, then at the end of the
voting process, the wizar:
1. waits for the baking app to be opened again
2. restarts the baker process(es) involved