[Chore] Use python3 package in Ubuntu focal packages
Problem: in bd2ecfbf38b22ffd1f744b403e93317889c0ce0c we introduced
a lower bound of python to version 3.8 and to do so a specific
package is used instead of the generic python3 on both Ubuntu
bionic and focal.
This is however only necessary for bionic and it's causing problems
on Raspberry Pi OS installations, as there installing python3.8 is
not simple (and there is no need to, as the python3 package is for
3.9).

Solution: use the python3 package on both focal and jammy, only use
the python3.8 package on bionic.