[#642] Stop node service when removing its data
Problem: the setup wizard can remove (most of) the content of a
node data dir if the user chooses to do so instead of keeping it.
However, if there is a node service still running, this data dir
can be partially re-filled while the wizard is still running and
so cause issues, because some node commands need the data dir to
be clean.

Solution: Stop the node service in the wizard before cleaning its
data directory.