[#644] Allow setup to resume snapshot download
Problem: Currently, if node snapshot download fails (e.g. for network
issues)
this step fails and the user needs to download a snapshot again.
This can be very inconvenient as some snapshots are very large
and downloading them can take a long time.

Solution: Use `wget` with the `--continue` option to be able to resume
download.