[#705] Better show the error logs of the wizards
Problem: the location of the log file is unclear from
the message given by the wizards. They also don't attempt
to show the logs or suggest how to do it.

Solution: Show last 10 lines of the error output. Suggest how to see
the full log. Separate log entries in the file by the dates.