[#782] Support tzinit provider
Problem: tzinit provider should be supported as
it is seem to be the most stable in the future.
Unfortunately, it does not provide as much metadata
as previous providers, thus sha256 is impossible to
retrieve, but filesize could be taken from the http
response itself.

Solution: Add tzinit provider and make it conform
snapshot metadata format used throught the codebase.